REP-10 Fix Replication query field name conflict
The encompassing field for replication queries and the field for the
configuration objects we've called "replications" had the same name.
This caused a bug where any reference to a replication was assumed to be
the start of a new replication query. This commit changes the name of a
replication configuration object to "repsync" to fix the issue.