REP-36 Add functionality to convert old configs and sites
Turns the MetacardConfigLoader into the legacyDataConverter, a bean that
will look for outdated data and convert it into the current version.It
will find metacard replication configs saved in the catalog, convert them into the
latest version, and move them into the ReplicationPersistentStore. It
will also take old sites that were saved under the "replication"
persistence type and save them with the "replication_site" persistence
type.