Fix E2E update test bug
The syncUpdate test was comparing metacard modified time to see if a
product was updated. When replicating a create, the metacard on the destination node gets the same
metacard.modified date as on the source node. When replicating
an update, however, the metacard on the destination node gets a metacard.modified date of its
local system time. This was causing the new metacard.modified date to be
earlier than the previous one, while the tests were expecting it to be
after.